RotiPortal/backend/requirements.txt
Flask==2.2.5
google-cloud-firestore==2.18.0
Flask-Cors==4.0.0